requirements.txt
torch
transformers
pillow
requests
open_clip-torch
ftfy
huggingface-hub
scipy
fastapi
uvicorn[standard]
jinja2
sqlalchemy
databases
asyncpg
pyjwt
apscheduler
httpx
python-dotenv
rake-nltk
pyyaml
pandas
psycopg
timescale_vector
beautifulsoup4
datasets
alembic
pgvector
symspellpy
openai
tenacity>=8.2.0
einops
timm
google-cloud-aiplatform
google-genai
python-dotenv
